feat: add support for Authelia, refactor OIDC
