+++ internal/auth/oidc.go
@@ -1,0 +1,6 @@
+package auth
+
+const AUTHELIA = "authelia"
+const AUTHENTIK = "authentik"
+const KEYCLOAK = "keycloak"
+const ZITADEL = "zitadel"
