+++ internal/controllers/webserver/handlers/auth_settings.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/invopop/ctxi18n/i18n"
 	"github.com/labstack/echo/v4"
+	"github.com/open-uem/openuem-console/internal/auth"
 	"github.com/open-uem/openuem-console/internal/views/admin_views"
 	"github.com/open-uem/openuem-console/internal/views/partials"
 )
@@ -24,7 +25,6 @@ func (h *Handler) AuthenticationSettings(c echo.Context) error {
 		oidcServer := c.FormValue("authentication-oidc-server")
 		oidcClientID := c.FormValue("authentication-oidc-client-id")
 		oidcRole := c.FormValue("authentication-oidc-role")
-		oidcKeycloakPublicKey := c.FormValue("authentication-oidc-keycloak-public-key")
 
 		useCertificates, err := strconv.ParseBool(c.FormValue("authentication-use-certificates"))
 		if err != nil {
@@ -59,7 +59,6 @@ func (h *Handler) AuthenticationSettings(c echo.Context) error {
 			oidcServer = ""
 			oidcClientID = ""
 			oidcRole = ""
-			oidcKeycloakPublicKey = ""
 		}
 
 		autoCreate, err := strconv.ParseBool(c.FormValue("authentication-oidc-auto-create"))
@@ -72,7 +71,7 @@ func (h *Handler) AuthenticationSettings(c echo.Context) error {
 			return RenderError(c, partials.ErrorMessage(i18n.T(c.Request().Context(), "authentication.could_not_parse_oidc_auto_approve"), true))
 		}
 
-		allowedProviders := []string{"authentik", "keycloak", "zitadel"}
+		allowedProviders := []string{auth.AUTHELIA, auth.AUTHENTIK, auth.KEYCLOAK, auth.ZITADEL}
 		if useOIDC && (oidcProvider == "" || (oidcProvider != "" && !slices.Contains(allowedProviders, oidcProvider))) {
 			return RenderError(c, partials.ErrorMessage(i18n.T(c.Request().Context(), "authentication.provider_not_valid"), true))
 		}
@@ -89,11 +88,7 @@ func (h *Handler) AuthenticationSettings(c echo.Context) error {
 			return RenderError(c, partials.ErrorMessage(i18n.T(c.Request().Context(), "authentication.role_required"), true))
 		}
 
-		if useOIDC && oidcRole != "" && oidcProvider == "keycloak" && oidcKeycloakPublicKey == "" {
-			return RenderError(c, partials.ErrorMessage(i18n.T(c.Request().Context(), "authentication.keycloak_public_key_required"), true))
-		}
-
-		if err := h.Model.SaveAuthenticationSettings(useCertificates, allowRegister, useOIDC, oidcProvider, oidcServer, oidcClientID, oidcRole, oidcKeycloakPublicKey, autoCreate, autoApprove); err != nil {
+		if err := h.Model.SaveAuthenticationSettings(useCertificates, allowRegister, useOIDC, oidcProvider, oidcServer, oidcClientID, oidcRole, autoCreate, autoApprove); err != nil {
 			return RenderError(c, partials.ErrorMessage(i18n.T(c.Request().Context(), "authentication.settings_not_saved", err.Error()), true))
 		}
